Tests/LibWeb/Text/expected/css/webkit-keyframes-alias.txt

Rule type: 7
Rule is CSSKeyframesRule: true
Animation name: test-animation
Number of keyframes: 2
